Move shared state synchronization to its own module.
This makes it impossible for other modules to access the underlying
RwLock directly. The _only_ way to obtain write access to the shared
state is to use the KeystoreSharedStateWriteGuard which forces updates
to go through an atomic storage transaction.
