+++ src/lib.rs
@@ -32,6 +32,7 @@ pub mod persistence;
 pub mod reader;
 mod secret;
 pub mod sparse_merkle_tree;
+mod state;
 #[cfg(any(test, bench, feature = "testing"))]
 pub mod testing;
 pub mod transactions;
@@ -54,11 +55,14 @@ use crate::{
     key_scan::{receive_history_entry, BackgroundKeyScan, ScanOutputs},
     loader::KeystoreLoader,
     persistence::AtomicKeystoreStorage,
+    state::{
+        KeystoreSharedState, KeystoreSharedStateReadGuard, KeystoreSharedStateRwLock,
+        KeystoreSharedStateWriteGuard,
+    },
     txn_builder::*,
 };
 use arbitrary::Arbitrary;
 use async_scoped::AsyncScope;
-use async_std::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use async_std::task::sleep;
 use async_trait::async_trait;
 use atomic_store::{
@@ -67,7 +71,7 @@ use atomic_store::{
 };
 use core::fmt::Debug;
 use espresso_macros::ser_test;
-use futures::{channel::oneshot, future::TryFuture, prelude::*, stream::Stream};
+use futures::{channel::oneshot, prelude::*, stream::Stream};
 use jf_cap::{
     errors::TxnApiError,
     freeze::FreezeNote,
@@ -1474,7 +1478,7 @@ pub struct Keystore<
     //  * promise completion handles for futures returned by sync(), indexed by the timestamp at
     //    which the corresponding future is supposed to complete. Handles are added in sync() (main
     //    thread) and removed and completed in the event thread
-    mutex: Arc<RwLock<KeystoreSharedState<'a, L, Backend, Meta>>>,
+    mutex: Arc<KeystoreSharedStateRwLock<'a, L, Backend, Meta>>,
     // Handle for the background tasks running the event handling loop and retroactive ledger scans.
     // When dropped, this handle will cancel the tasks.
     task_scope: AsyncScope<'a, ()>,
@@ -1527,222 +1531,6 @@ impl<T: Serialize + DeserializeOwned> LoadStore for EncryptingResourceAdapter<T>
                     inner: Box::new(err),
                 })?;
         bincode::serialize(&ciphertext).map_err(|source| ASPersistenceError::BincodeSer { source })
-    }
-}
-
-/// Keystore state which is shared with event handling threads.
-pub struct KeystoreSharedState<
-    'a,
-    L: Ledger,
-    Backend: KeystoreBackend<'a, L>,
-    Meta: Serialize + DeserializeOwned + Send,
-> {
-    state: KeystoreState<'a, L>,
-    model: KeystoreModel<'a, L, Backend, Meta>,
-    sync_handles: Vec<(EventIndex, oneshot::Sender<()>)>,
-    txn_subscribers: HashMap<TransactionUID<L>, Vec<oneshot::Sender<TransactionStatus>>>,
-    pending_foreign_txns: HashMap<Nullifier, Vec<oneshot::Sender<TransactionStatus>>>,
-    pending_key_scans: HashMap<UserAddress, Vec<oneshot::Sender<()>>>,
-}
-
-impl<'a, L: Ledger, Backend: KeystoreBackend<'a, L>, Meta: Serialize + DeserializeOwned + Send>
-    KeystoreSharedState<'a, L, Backend, Meta>
-{
-    async fn commit(&mut self) -> Result<(), KeystoreError<L>> {
-        self.model.persistence.commit().await;
-        self.model.assets.commit()?;
-        self.model
-            .atomic_store
-            .commit_version()
-            .map_err(KeystoreError::from)
-    }
-
-    async fn revert(&mut self) -> Result<(), KeystoreError<L>> {
-        self.model.assets.revert()?;
-        self.model.persistence.revert().await;
-        // Reload in-memory state after the revert.
-        self.state = self.model.persistence.load().await?;
-        Ok(())
-    }
-}
-
-impl<'a, L: Ledger, Backend: KeystoreBackend<'a, L>, Meta: Serialize + DeserializeOwned + Send>
-    KeystoreSharedState<'a, L, Backend, Meta>
-{
-    pub fn backend(&self) -> &Backend {
-        &self.model.backend
-    }
-
-    pub fn backend_mut(&mut self) -> &mut Backend {
-        &mut self.model.backend
-    }
-
-    pub fn state(&self) -> &KeystoreState<'a, L> {
-        &self.state
-    }
-
-    pub fn rng(&mut self) -> &mut ChaChaRng {
-        &mut self.model.rng
-    }
-}
-
-/// A guard for [KeystoreSharedState] that allows writing and gracefully handles errors.
-///
-/// Unlike [RwLockWriteGuard], [KeystoreSharedStateWriteGuard] does not dereference to the target,
-/// so the [KeystoreSharedState] cannot be freely edited directly. Instead,
-/// [KeystoreSharedStateWriteGuard] provides an [update](Self::update) function, which can be used
-/// to apply a closure to the shared state. The guard checks the result of this closure and takes
-/// note if it fails. Upon being dropped, if any operation on the guard fails, the persistent _and_
-/// in-memory state will be reverted to the state at the time the guard was created. All operations
-/// between creating and dropping the guard are atomic.
-///
-/// This abstraction simplifies the problem of reverting previous changes when a later change fails,
-/// and avoids confusion about when to open a storage transaction and when to commit it. With
-/// [KeystoreSharedStateWriteGuard], both of these questions have simple answers: you open a
-/// transaction when you obtain write access to the resource and you keep it open until you drop
-/// your write access. All changes made within a transaction are reverted atomically and
-/// automatically if the transaction fails.
-pub struct KeystoreSharedStateWriteGuard<
-    'l,
-    'a,
-    L: Ledger,
-    Backend: KeystoreBackend<'a, L>,
-    Meta: Send + Serialize + DeserializeOwned,
-> {
-    guard: RwLockWriteGuard<'l, KeystoreSharedState<'a, L, Backend, Meta>>,
-    failed: bool,
-}
-
-impl<
-        'l,
-        'a,
-        L: Ledger,
-        Backend: KeystoreBackend<'a, L>,
-        Meta: Send + Serialize + DeserializeOwned,
-    > KeystoreSharedStateWriteGuard<'l, 'a, L, Backend, Meta>
-{
-    async fn new(
-        mutex: &'l RwLock<KeystoreSharedState<'a, L, Backend, Meta>>,
-    ) -> KeystoreSharedStateWriteGuard<'l, 'a, L, Backend, Meta> {
-        Self {
-            guard: mutex.write().await,
-            failed: false,
-        }
-    }
-
-    /// Mutate the [KeystoreSharedState] with a closure.
-    ///
-    /// `op` defines a self-contained operation to apply to the shared state. If it fails, or if any
-    /// later operation on the same guard fails, all of its changes will be reverted. Otherwise, its
-    /// changes will be committed when the guard is dropped.
-    ///
-    /// The result of [update](Self::update) is the result of `op`.
-    ///
-    /// # Lifetimes
-    ///
-    /// The lifetime `'s` deserves some discussion. It is more idiomatic for closure parameters that
-    /// take a reference to use a higher-rank type bound, like
-    ///
-    /// ```ignore
-    /// for<'s> FnOnce(&'s mut T) -> BoxFuture<'s, R>
-    /// ```
-    ///
-    /// However, this simpler type signature states something that we don't want to be true: that
-    /// the function parameter, `op`, returns a future that _only_ borrows from its argument
-    /// (indicated by the polymorphic `'s`). We want to allow closures that borrow from _both_ their
-    /// argument and their environment. Without this, it is impossible to write a function that
-    /// takes a reference parameter and uses that parameter during an [update](Self::update) (at
-    /// least, not without cloning every reference parameter and moving the clone into the closure).
-    /// Unfortunately, there is not a good way to require that the future is bounded by two
-    /// lifetimes. We'd like to write something like
-    ///
-    /// ```ignore
-    /// F: for<'a> FnOnce(&'a T) -> BoxFuture<'a + 'b, R>
-    /// ```
-    ///
-    /// or
-    ///
-    /// ```ignore
-    /// for<'a>
-    ///     F: FnOnce(&'a T) -> Fut,
-    ///     Fut: 'a,
-    ///     Fut: 'b,
-    /// ```
-    ///
-    /// But neither of these can be expressed in the Rust type system.
-    ///
-    /// So, instead of requiring a closure that works for _every_ `'s`, we let the caller provide a
-    /// lifetime `'s`, which can be the lifetime of some reference that outlives the
-    /// [update](Self::update) operation. This is why the parameter `&'s mut self` is crucial: it
-    /// allows the closure `op` to _also_ borrow from its argument, since it's argument, the shared
-    /// state managed by this guard, is also borrowed from `self` and thus has the same lifetime,
-    /// `'s`.
-    ///
-    /// This is actually the reason for this guard type's existence. Were it not for this limitation
-    /// of the type system, we could simply write an update function that takes a
-    /// `RwLock<KeystoreSharedState<'a, L, Backend, Meta>>`, locks it, and then calls `op` on the
-    /// temporary guard. This fails because the lifetime of the guard itself must be the same as the
-    /// lifetime of the captured environment chosen by the caller, and so the guard must outlive the
-    /// update function; hence, we define this [KeystoreSharedStateWriteGuard] type, which can live
-    /// on the caller's stack for the duration of the [update](Self::update), allowing the caller to
-    /// name its lifetime: `'s`.
-    ///
-    /// # Errors
-    ///
-    /// If `op` returns an error, the error propagates out of [update](Self::update), and the guard
-    /// enters a failed state. When it is dropped, any changes made by `op` will be reverted.
-    ///
-    /// If a previous [update](Self::update) failed, subsequent calls to [update](Self::update) will
-    /// fail immediately without invoking `op` at all.
-    pub async fn update<'s, F, Fut>(&'s mut self, op: F) -> Result<Fut::Ok, Fut::Error>
-    where
-        F: FnOnce(&'s mut KeystoreSharedState<'a, L, Backend, Meta>) -> Fut,
-        Fut: TryFuture<Error = KeystoreError<L>>,
-    {
-        if self.failed {
-            return Err(KeystoreError::Failed {
-                msg: "calling update on a transaction that has already failed".into(),
-            });
-        }
-
-        let failed = &mut self.failed;
-        op(&mut *self.guard)
-            .inspect_err(|_| {
-                // Enter the failed state if the operation fails.
-                //
-                // The reason we do things this way, instead of just reverting the change right
-                // here, is a bit silly. In order to give a name to the lifetime of `op`s argument,
-                // we defined the lifetime parameter `'s`, which necessarily outlives the body of
-                // this function. Since the future returned by `op` borrows from `self.guard` with
-                // lifetime `'s`, the compiler believes that `self.guard` is borrowed mutably until
-                // after this function returns, even though this is impossible, since we know that
-                // the future will be driven to completion before this error handler gets called.
-                // Nevertheless, entering a failure state and then cleaning things up on drop (after
-                // this function returns and the compiler realizes `self.guard` can no longer be
-                // borrowed) is an acceptable workaround.
-                *failed = true;
-            })
-            .into_future()
-            .await
-    }
-}
-
-impl<
-        'l,
-        'a,
-        L: Ledger,
-        Backend: KeystoreBackend<'a, L>,
-        Meta: Send + Serialize + DeserializeOwned,
-    > Drop for KeystoreSharedStateWriteGuard<'l, 'a, L, Backend, Meta>
-{
-    fn drop(&mut self) {
-        async_std::task::block_on(async move {
-            if self.failed {
-                self.guard.revert().await.unwrap();
-            } else {
-                self.guard.commit().await.unwrap();
-            }
-        });
     }
 }
 
@@ -1868,25 +1656,14 @@ impl<
             _marker: Default::default(),
             _marker2: Default::default(),
         };
-
         // Ensure the native asset type is always recognized.
         model.create_native_asset(None)?;
 
-        let sync_handles = Vec::new();
-        let txn_subscribers = HashMap::new();
-        let pending_foreign_txns = HashMap::new();
-        let mutex = Arc::new(RwLock::new(KeystoreSharedState {
+        let mutex = Arc::new(KeystoreSharedStateRwLock::new(
             state,
             model,
-            sync_handles,
-            txn_subscribers,
-            pending_foreign_txns,
-            pending_key_scans: key_scans
-                .iter()
-                .map(|(key, _)| (key.clone(), vec![]))
-                .collect(),
-        }));
-
+            key_scans.iter().map(|(key, _)| key.clone()),
+        ));
         let mut scope = unsafe {
             // Creating an AsyncScope is considered unsafe because `std::mem::forget` is allowed
             // in safe code, and forgetting an AsyncScope can allow its inner futures to
@@ -1908,7 +1685,8 @@ impl<
                 async move {
                     let mut foreign_txns_awaiting_memos = HashMap::new();
                     while let Some((event, source)) = events.next().await {
-                        while let Err(err) = KeystoreSharedStateWriteGuard::new(&mutex)
+                        while let Err(err) = mutex
+                            .write()
                             .await
                             .update(|state| {
                                 update_ledger(
@@ -1948,11 +1726,11 @@ impl<
 
     /// Access the shared state directly.
     pub async fn write(&self) -> KeystoreSharedStateWriteGuard<'_, 'a, L, Backend, Meta> {
-        KeystoreSharedStateWriteGuard::new(&self.mutex).await
+        self.mutex.write().await
     }
 
     /// Access the shared state directly.
-    pub async fn read(&self) -> RwLockReadGuard<'_, KeystoreSharedState<'a, L, Backend, Meta>> {
+    pub async fn read(&self) -> KeystoreSharedStateReadGuard<'_, 'a, L, Backend, Meta> {
         self.mutex.read().await
     }
 
@@ -2870,7 +2648,8 @@ impl<
                 while !finished {
                     let (next_event, source) = events.next().await.unwrap();
                     loop {
-                        match KeystoreSharedStateWriteGuard::new(&mutex)
+                        match mutex
+                            .write()
                             .await
                             .update(|state| {
                                 update_key_scan(&address, next_event.clone(), source, state).boxed()
